tests/wpt/tests/IndexedDB/idbdatabase_createObjectStore4.htm

<!DOCTYPE html>
<title>IDBDatabase.createObjectStore() - attempt to create an object store that already exists </title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/support.js"></script>

<script>

var t = async_test(),
    open_rq = createdb(t)

open_rq.onupgradeneeded = function(e) {
    var db = e.target.result
    db.createObjectStore("dupe")
    assert_throws_dom(
        'ConstraintError',
        function() { db.createObjectStore("dupe") })

    // Bonus test creating a new objectstore after the exception
    db.createObjectStore("dupe ")
    t.done()
}

</script>

<div id="log"></div>
